Improve Windows compatibility
Add command to Dockerfile to automatically remove any CRLF line termination from the entry point script. They are sometimes added by git when it runs on Windows and they prevent the script from running when the file is copied back on a Linux image.
Add code in cmqttd.py that is triggered when the Python code is run directly on Windows. It deals with an issue related to windows ports' behavior within the event loop that results in an error. The issue is documented here: Configuration of windows event loop for libraries python/cpython#81554
It is useful when porting the python code to a Windows system.